Move timesyncd API to /io/hass/os/Config/Timesyncd DBus path
Per discussion in review, move the new API to path that clearer
represents this API sets system(d) configuration. The Config path can
later contain more toggles setting persistent OS configuration.
